Report fields are lowercase, and the integer value is renamed to "int64Value"

+++ aggregator/aggregator_test.go
@@ -42,7 +42,7 @@ func TestNewAggregator(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 			Labels: map[string]string{
 				"key": "value1",
@@ -54,7 +54,7 @@ func TestNewAggregator(t *testing.T) {
 			StartTime: time.Unix(10, 0),
 			EndTime:   time.Unix(11, 0),
 			Value: metrics.MetricValue{
-				IntValue: 333,
+				Int64Value: 333,
 			},
 			Labels: map[string]string{
 				"key": "value2",
@@ -66,7 +66,7 @@ func TestNewAggregator(t *testing.T) {
 			StartTime: time.Unix(100, 0),
 			EndTime:   time.Unix(110, 0),
 			Value: metrics.MetricValue{
-				IntValue: 555,
+				Int64Value: 555,
 			},
 			Labels: map[string]string{
 				"key": "value3",
@@ -171,7 +171,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -186,7 +186,7 @@ func TestAggregator_AddReport(t *testing.T) {
 				StartTime: time.Unix(0, 0),
 				EndTime:   time.Unix(1, 0),
 				Value: metrics.MetricValue{
-					IntValue: 10,
+					Int64Value: 10,
 				},
 			},
 		}
@@ -209,7 +209,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -219,7 +219,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(2, 0),
 			EndTime:   time.Unix(3, 0),
 			Value: metrics.MetricValue{
-				IntValue: 5,
+				Int64Value: 5,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -234,7 +234,7 @@ func TestAggregator_AddReport(t *testing.T) {
 				StartTime: time.Unix(0, 0),
 				EndTime:   time.Unix(3, 0),
 				Value: metrics.MetricValue{
-					IntValue: 15,
+					Int64Value: 15,
 				},
 			},
 		}
@@ -260,7 +260,7 @@ func TestAggregator_AddReport(t *testing.T) {
 				"key1": "value1",
 			},
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -273,7 +273,7 @@ func TestAggregator_AddReport(t *testing.T) {
 				"key1": "value2",
 			},
 			Value: metrics.MetricValue{
-				IntValue: 5,
+				Int64Value: 5,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -291,7 +291,7 @@ func TestAggregator_AddReport(t *testing.T) {
 					"key1": "value1",
 				},
 				Value: metrics.MetricValue{
-					IntValue: 10,
+					Int64Value: 10,
 				},
 			},
 			{
@@ -302,7 +302,7 @@ func TestAggregator_AddReport(t *testing.T) {
 					"key1": "value2",
 				},
 				Value: metrics.MetricValue{
-					IntValue: 5,
+					Int64Value: 5,
 				},
 			},
 		}
@@ -328,7 +328,7 @@ func TestAggregator_AddReport(t *testing.T) {
 				"key1": "value1",
 			},
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err == nil || !strings.Contains(err.Error(), "StartTime > EndTime") {
 			t.Fatalf("Expected error containing \"StartTime > EndTime\", got: %+v", err.Error())
@@ -347,7 +347,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -357,7 +357,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(2, 0),
 			Value: metrics.MetricValue{
-				IntValue: 5,
+				Int64Value: 5,
 			},
 		}); err == nil || !strings.Contains(err.Error(), "time conflict") {
 			t.Fatalf("Expected error containing \"time conflict\", got: %+v", err)
@@ -376,7 +376,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -386,7 +386,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(2, 0),
 			EndTime:   time.Unix(3, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 			Labels: map[string]string{
 				"foo": "bar",
@@ -409,7 +409,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(4, 0),
 			EndTime:   time.Unix(5, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
@@ -419,7 +419,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(6, 0),
 			EndTime:   time.Unix(7, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 			Labels: map[string]string{
 				"foo": "bar",
@@ -450,7 +450,7 @@ func TestAggregator_AddReport(t *testing.T) {
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 			Value: metrics.MetricValue{
-				IntValue: 10,
+				Int64Value: 10,
 			},
 		}); err != nil {
 			t.Fatalf("Unexpected error when adding report: %+v", err)
