+++ sender/retry_test.go
@@ -36,7 +36,7 @@ func TestRetryingSender(t *testing.T) {
 		Id: "report1",
 		MetricReport: metrics.MetricReport{
 			Name:      "int-metric",
-			Value:     metrics.MetricValue{IntValue: 10},
+			Value:     metrics.MetricValue{Int64Value: 10},
 			StartTime: time.Unix(0, 0),
 			EndTime:   time.Unix(1, 0),
 		},
@@ -45,7 +45,7 @@ func TestRetryingSender(t *testing.T) {
 		Id: "report2",
 		MetricReport: metrics.MetricReport{
 			Name:      "int-metric",
-			Value:     metrics.MetricValue{IntValue: 30},
+			Value:     metrics.MetricValue{Int64Value: 30},
 			StartTime: time.Unix(10, 0),
 			EndTime:   time.Unix(11, 0),
 		},
@@ -54,7 +54,7 @@ func TestRetryingSender(t *testing.T) {
 		Id: "report3",
 		MetricReport: metrics.MetricReport{
 			Name:      "int-metric",
-			Value:     metrics.MetricValue{IntValue: 30},
+			Value:     metrics.MetricValue{Int64Value: 30},
 			StartTime: time.Unix(20, 0),
 			EndTime:   time.Unix(21, 0),
 		},
